Move hello_quantum to its own package
